Adding our own html
I added a folder html were there are pages to use as a front of our server. They should be enough to test all 3 GET, POST and DELETE methods. I also tweaked two source code files so that this is the default version. I did not touch the configs yet, so those are still using the files from www folder